backend/requirements.txt
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
pydantic[email]>=2.6.0,<3.0.0
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.12.0,<2.0.0
python-multipart>=0.0.6,<1.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-dotenv>=1.0.0,<2.0.0
pandas>=2.0.0,<3.0.0
openpyxl>=3.1.0,<4.0.0
google-api-python-client>=2.100.0,<3.0.0
google-auth-httplib2>=0.1.0,<1.0.0
google-auth-oauthlib>=1.1.0,<2.0.0
httpx>=0.25.0,<1.0.0
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
email-validator>=2.0.0,<3.0.0 